+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 * (BREAKING) Drop gradle 7.0 support (7.1 min requirement due to spotbugs plugin)
-* Update checkstyle 10.12.7 -> 10.26.1 (java 21 support)
+* Update checkstyle 10.12.7 -> 10.26.1 (java 21 records support)
     - Add [PatternVariableAssignment](https://checkstyle.sourceforge.io/checks/coding/patternvariableassignment.html)
     - Add [UnnecessaryNullCheckWithInstanceOf](https://checkstyle.sourceforge.io/checks/coding/unnecessarynullcheckwithinstanceof.html)
     - Add [ConstructorsDeclarationGrouping](https://checkstyle.sourceforge.io/checks/coding/constructorsdeclarationgrouping.html)
@@ -8,7 +8,7 @@
       * On java 8 checkstyle plugin simply isn't applied
 * Update spotbugs 4.8.3 -> 4.9.3 (requires java 11)
     - On java 8 version 4.8.6 would be selected (the latest compatible with java 8)
-* Update pmd 6.55 -> 7.15.0 (java 21, 22 support)
+* Update pmd 6.55 -> 7.16.0 (java 25 support)
     - Fix gradle plugin for pmd 7 (pmd 7 split jar so plugin classpath must be overridden )
     - Fix default pmd config: remove legacy rules exclusions
 * Update codenarc 3.4.0 -> 3.6.0